- Improved file lookup to btree lookup. - Now the partial lookup uses btree lookup to find a particular file/folder - Two functions ntfs_dir_open_meta_inner_btree and ntfs_find_idxentry do the job. - Updated .h to take care of warning

+++ tsk3/fs/tsk_ntfs.h
@@ -445,7 +445,7 @@ extern "C" {
  * 8 bytes of the structure 
  */
 #define GET_IDXENTRY_SUB(fs, e)	\
-	(tsk_getu64(fs->endian, (int)e + tsk_getu16(fs->endian, e->idxlen) - 8))
+	(tsk_getu64(fs->endian, (uint64_t)e + tsk_getu16(fs->endian, e->idxlen) - 8))
